MOSIP-13522

+++ registration-processor/registration-processor-registration-status-service-impl/src/main/java/io/mosip/registration/processor/status/service/RegistrationStatusService.java
@@ -139,7 +139,7 @@ public interface RegistrationStatusService<T, U, D> {
 	 *
 	 * @return the InProgress packets count
 	 */
-	public Integer getInProgressPacketsCount();
+	public long getInProgressPacketsCount();
 
 	/**
 	 * Check Rid if uin is available.
